DNS M4 deferral closeout: prune 32 unwired metrics, clippy clean, benchmark baseline, profile audit
8-workstream closeout of DNS Milestone 4 deferred items. Status: Closed with
accepted deferrals (external live-wire interop, external DNSSEC tooling,
remote CI status visibility).

Workstream 1 — Remote CI: Documented status-visibility limitation (connector
does not surface direct-push workflow runs). Local release gate is canonical
release signal until alternative visibility path is established.

Workstream 2 — External live-wire interop: Deferred. conformance.sh already
distinguishes internal in-process interop (7 suites, in CI) from optional
external tool checks (operator-validated).

Workstream 3 — Unwired metrics: 32 record_* methods + 32 backing fields +
20 DnsMetricsSummary fields deleted from metrics.rs (1128 -> 504 lines).
Kept 17 production-active methods, all wired to real call sites. 5 watchable
metrics (dns_active_tcp_connections, dns_recursive_circuit_breaker_opens_total,
dns_zone_reload_failures_total, dns_encode_failures_total,
dns_dnssec_signing_failures_total) continue to emit via direct metrics::counter!
in production code paths.

Workstream 4 — Benchmark baseline: Local reference at
benchmarks/dns/results/2026-07-07-baseline.md (commit 4a76cc74, i9-9900K,
rustc 1.95.0, 53 criterion timings). Benchmarks remain reference-only.

Workstream 5 — Production profiles: Production-Supported Boundary section,
Release Support Matrix (8 profiles x 4 cols), DNSSEC Coverage Boundary
sub-section, encrypted/transfer profile scope clarifications.

Workstream 6 — External DNSSEC tooling: Deferred. DNSSEC profile split into
known-vector / live-signed / external-tooling tiers; the first two are
covered, the third is deferred.

Workstream 7 — Scripts: All 5 DNS scripts pass bash -n.

Workstream 8 — Closeout record: plans/dns_milestone_4_deferred_items_closeout_complete.md

Also brought cargo clippy -p synvoid-dns --all-targets -- -D warnings from
63 errors to clean (10 #[allow(too_many_arguments)] for genuine large-fn
sites; rest are real fixes across src/, tests/, benches/).

Verification:
  cargo fmt --all -- --check                              PASS
  cargo clippy -p synvoid-dns --all-targets -- -D warnings  PASS
  cargo test -p synvoid-dns --release --no-fail-fast       PASS (1101 / 31 suites)
  cargo check -p synvoid-dns --all-features                PASS
  cargo bench -p synvoid-dns --no-run                      PASS (5 suites)
  bash -n (5 DNS scripts)                                  PASS

Total: 1101 DNS tests passing in release mode.

+++ crates/synvoid-dns/tests/dns_interop_authoritative.rs
@@ -328,7 +328,7 @@ fn decode_wire_name(resp: &[u8], start: usize) -> Option<String> {
             if pos + 1 >= resp.len() {
                 return None;
             }
-            let offset = ((len & 0x3F) as usize) << 8 | resp[pos + 1] as usize;
+            let offset = (len & 0x3F) << 8 | resp[pos + 1] as usize;
             if !jumped {
                 jumped = true;
             }
